requirements.txt
annotated-doc==0.0.4
asgiref==3.11.1
click==8.3.1
diff-match-patch==20241021
Django==6.0.2
django-import-export==4.4.0
django-pagination==1.0.7
django-typer==3.6.3
django-unfold==0.82.0
et_xmlfile==2.0.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
pillow==12.1.1
psycopg2-binary==2.9.11
Pygments==2.19.2
python-dateutil==2.9.0.post0
rich==14.3.3
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
typer==0.24.1
